it didnt work, it works only with 777 access right for voices dir. whisper models dir idea wasnt realized properly, one more try

+++ handlers.py
@@ -12,7 +12,7 @@ MAX_VOICE_DURATION = 60
 pool = ThreadPoolExecutor(max_workers=1) 
 router = Router()
 
-logger = logging.getLogger(__name__) # __name__ автоматически даст имя модуля: 'fast_whisp_test'
+logger = logging.getLogger(__name__) 
 
 
 @router.message(CommandStart())
@@ -34,9 +34,6 @@ async def transcribe(message: Message):
 
     voice = message.reply_to_message.voice
 
-    # if voice.duration > MAX_VOICE_DURATION:
-    #     await message.reply_to_message.reply(f'гс должно быть не более {MAX_VOICE_DURATION} сек')
-    
     if not voice:
         await message.answer('нужно ответить именно на ГС')
         return
@@ -63,9 +60,11 @@ async def transcribe(message: Message):
         loop = asyncio.get_running_loop()
         result = await loop.run_in_executor(pool, processor.transcribe, file_id)
         await message.reply_to_message.reply("".join(result))
-        os.remove(f'voices/{file_id}.ogg')
     except Exception as e:
         logger.error(f'AN ERROR OCCURED WHILE TRANCRIBITION: {e}')
         await message.reply_to_message.reply('ой, что-то пошло не так(\nприносим извинения за временные неудобства\nповторите попытку чуть позже')
+    finally:
+        if os.path.exists(f'voices/{file_id}.ogg'):
+            os.remove(f'voices/{file_id}.ogg')
